% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{appc_clean_data_directory}
\alias{appc_clean_data_directory}
\title{delete all installed data files in the user's data directory for the \code{appc} package}
\usage{
appc_clean_data_directory()
}
\description{
delete all installed data files in the user's data directory for the \code{appc} package
}
