% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{contiguous_us}
\alias{contiguous_us}
\title{return the s2 shape of the 2020 contiguous united states}
\usage{
contiguous_us()
}
\value{
s2_geography object
}
\description{
return the s2 shape of the 2020 contiguous united states
}
\examples{
contiguous_us()
}
