% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_smoke.R
\name{download_daily_smoke_data}
\alias{download_daily_smoke_data}
\title{download_daily_smoke_data(as.Date("2024-04-21"))}
\usage{
download_daily_smoke_data(date)
}
\description{
download_daily_smoke_data(as.Date("2024-04-21"))
}
