% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{get_census_tract_id}
\alias{get_census_tract_id}
\title{Get census tract identifier of \emph{the closest} census tract}
\usage{
get_census_tract_id(x, year)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{year}{a numeric data year passed to tigris to get state and tract boundaries}
}
\description{
Get census tract identifier of \emph{the closest} census tract
}
\details{
\code{tigris::tracts()} powers this, so set \code{options(tigris_use_cache = TRUE)}
to benefit from its caching.
According to \url{https://github.com/walkerke/tigris}, available years for tracts
and states are 1990, 2000, 2010 - 2022
}
\examples{
\dontrun{
get_census_tract_id(s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")), year = "2020")
}
}
