% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_smoke.R
\name{get_daily_smoke}
\alias{get_daily_smoke}
\title{get daily smoke data for a vector of s2 locations}
\usage{
get_daily_smoke(x, date)
}
\description{
get daily smoke data for a vector of s2 locations
}
\examples{
get_daily_smoke(
  x = s2::as_s2_cell(c("8841b39a7c46e25f", "8841a45555555555")),
  date = as.Date("2023-06-10")
)
}
