% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{get_elevation_summary}
\alias{get_elevation_summary}
\title{get elevation summary data}
\usage{
get_elevation_summary(x, fun = stats::median, buffer = 800)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{fun}{function to summarize extracted data}

\item{buffer}{distance from s2 cell (in meters) to summarize data}
}
\value{
a numeric vector of elevation summaries, the same length as \code{x}
}
\description{
get elevation summary data
}
\examples{
\dontrun{
get_elevation_summary(s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")))
}
}
