% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet.R
\name{get_gridmet_data}
\alias{get_gridmet_data}
\alias{install_gridmet_data}
\title{Get gridMET surface meteorological data}
\usage{
get_gridmet_data(x, dates, gridmet_var = c("tmmx", "tmmn", "pr", "rmax"))

install_gridmet_data(
  gridmet_var = c("tmmx", "tmmn", "pr", "rmax"),
  gridmet_year = as.character(1979:format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{dates}{a list of date vectors for the NARR data, must be the same length as \code{x}}

\item{gridmet_var}{a character string that is the name of a gridMET variable}

\item{gridmet_year}{a character string that is the year for the gridMET data; see details}
}
\value{
for \code{get_gridmet_data()}, a list of numeric vectors of gridMET values (the same length as \code{x} and \code{dates})

for \code{install_gridmet_data()}, a character string path to gridMET raster data
}
\description{
Daily, high spatial resolution (~4-km) data comes from the \href{https://www.climatologylab.org/gridmet.html}{Climatology Lab}
and is available for the contiguous US from 1979-yesterday.
}
\details{
GRIDMET data comes as 1/24th degree gridded data, which is about 4 sq km resolution. s2 geohashes
are intersected with this grid for matching with daily weather values.

gridMET variables are named:

\if{html}{\out{<div class="sourceCode">}}\preformatted{gridmet_variable <- c(
  temperature_max = "tmmx",
  temperature_min = "tmmn",
  precipitation = "pr",
  solar_radiation = "srad",
  wind_speed = "vs",
  wind_direction = "th",
  relative_humidity_max = "rmax",
  relative_humidity_min = "rmin"
  specific_humidity = "sph"
)
}\if{html}{\out{</div>}}
}
\examples{
d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2023-11-06")),
  "8841a45555555555" = as.Date(c("2023-06-22", "2023-08-15"))
)
get_gridmet_data(x = s2::as_s2_cell(names(d)), dates = d, gridmet_var = "tmmx")
get_gridmet_data(x = s2::as_s2_cell(names(d)), dates = d, gridmet_var = "pr")
}
\references{
\url{https://www.climatologylab.org/gridmet.html}

\url{https://www.northwestknowledge.net/metdata/data/}
}
