% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_smoke.R
\name{get_hms_smoke_data}
\alias{get_hms_smoke_data}
\alias{install_hms_smoke_data}
\title{Get smoke plume data from the NOAA's Hazard Mapping System}
\usage{
get_hms_smoke_data(x, dates, quiet = TRUE)

install_hms_smoke_data()
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object); currently required to be within the contiguous united states}

\item{dates}{a list of date vectors for the predictions, must be the same length as \code{x}}

\item{quiet}{silence progress messages?}
}
\value{
for \code{get_hms_smoke_data()}, a list of numeric vectors of smoke plume scores (the same length as \code{x} and \code{dates})

for \code{install_hms_smoke_data()}, a character string path to the installed RDS file
}
\description{
The HMS operates daily in near real-time by outlining the smoke polygon of each distinct smoke plume
and classifying it as "light", "medium", and "heavy". Since multiple plumes of varying or the same classification
can cover one another, the total smoke plume exposure is estimated as the weighted sum of all plumes, where
"light" = 1, "medium" = 2, and "heavy" = 3.
}
\details{
Daily HMS shapefiles are missing for 7 days within 2017-2023
("2017-04-27", "2017-05-31", "2017-06-01", "2017-06-01" "2017-06-22", "2017-11-12", "2018-12-31")
and will return zero values.  If files are available but no smoke plumes intersect, then a zero values is also returned.
}
\examples{
d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2017-11-06")),
  "8841a45555555555" = as.Date(c("2017-06-22", "2023-08-15"))
)
get_hms_smoke_data(x = s2::as_s2_cell(names(d)), dates = d)
}
\references{
\url{https://www.ospo.noaa.gov/Products/land/hms.html#about}
}
