% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra.R
\name{install_merra_data}
\alias{install_merra_data}
\alias{create_daily_merra_data}
\title{Installs MERRA PM2.5 data into user's data directory for the \code{appc} package}
\usage{
install_merra_data(merra_year = as.character(2016:2023))

create_daily_merra_data(date)
}
\arguments{
\item{date}{a date object that is the date for the merra data}

\item{year}{a year object that is the year for the merra data}
}
\value{
for \code{install_merra_data()}, a character string path to the merra data

for \code{get_daily_merra_data()}, a tibble with columns for s2,
date, and concentrations of PM2.5 total, dust, oc, bc, ss, so4
}
\description{
Installs MERRA PM2.5 data into user's data directory for the \code{appc} package

Downloads and computes MERRA PM2.5 data for a single day
}
