% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nei.R
\name{get_nei_point_summary}
\alias{get_nei_point_summary}
\title{get NEI point summary data}
\usage{
get_nei_point_summary(
  x,
  year,
  pollutant_code = c("PM25-PRI", "EC", "OC", "SO4", "NO3", "PMFINE"),
  buffer = 1000
)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{year}{a character string that is the year of the NEI data}

\item{pollutant_code}{the NEI pollutant to summarize}

\item{buffer}{distance from s2 cell (in meters) to summarize data}
}
\value{
a vector (the same length as \code{x}) of the sum of all point emissions within the buffer distance weighted
by the inverse of the distance squared to each emission point
}
\description{
get NEI point summary data
}
\examples{
\dontrun{
get_nei_point_summary(s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")), year = "2020")
}
}
