% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_smoke.R
\name{get_smoke_data}
\alias{get_smoke_data}
\title{Get smoke plume data from the NOAA's Hazard Mapping System}
\usage{
get_smoke_data(x, dates)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object); currently required to be within the contiguous united states}

\item{dates}{a list of date vectors for the predictions, must be the same length as \code{x}}
}
\value{
a list of numeric vectors of smoke plume scores (the same length as \code{x} and \code{dates})
}
\description{
The HMS operates daily in near real-time by outlining the smoke polygon of each distinct smoke plume
and classifying it as "light", "medium", and "heavy". Since multiple plumes of varying or the same classification
can cover one another, the total smoke plume exposure is estimated as the weighted sum of all plumes, where
"light" = 1, "medium" = 2, and "heavy" = 3.
}
\examples{
d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2023-11-06")),
  "8841a45555555555" = as.Date(c("2023-06-22", "2023-08-15"))
)
get_smoke_data(x = s2::as_s2_cell(names(d)), dates = d)
}
\references{
\url{https://www.ospo.noaa.gov/Products/land/hms.html#about}
}
