% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{get_traffic_summary}
\alias{get_traffic_summary}
\alias{install_traffic}
\title{Get traffic summary}
\usage{
get_traffic_summary(x, buffer = 400)

install_traffic(traffic_release = "hpms_2020_f12_aadt-2025-07-16")
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{buffer}{distance from s2 cell (in meters) to summarize data}

\item{traffic_release}{name of github release to download traffic data file from}
}
\value{
a list the same length as \code{x}, which each element having a numeric vector of
\code{aadtm_passenger}, \code{aadtm_trucks_buses}, \code{aadtm_tractor_trailer}
}
\description{
Highway Performance Monitoring System (HPMS) data from 2020 is summarized as the average daily total number of meters driven by
passenger vehicles, trucks/busses, and tractor-trailers on interstates, freeways, and expressways
within \code{buffer} meters of each s2 cell
}
\details{
Only roads with F_SYSTEM classification of 1 ("interstate") or 2 ("principal arterial - other freeways and expressways")
are used. Passenger vehicles (FHWA 1-3) are calculated as the total minus FHWA class 4-7 (single unit) and 8-13 (combo)
}
\examples{
get_traffic_summary(
  s2::as_s2_cell(c("8841b6abd8207619", "8841b4f6affffffb", "8841b39f07f7d899")))
\dontrun{
# randomly sample 100 level 18 cells from s2 level-9: 8841b4
# https://igorgatis.github.io/ws2/?cells=8841b4
# use their centroids as the level 30 s2 cells
set.seed(1)
my_s2_cells <-
  s2::s2_covering_cell_ids(s2::s2_cell_polygon(s2::as_s2_cell("8841b4")),
                           min_level = 18, max_level = 18) |>
  unlist()|>
  sample(size = 100)|>
  s2::s2_cell_center()|>
  s2::as_s2_cell()
get_traffic_summary(my_s2_cells) |>
  dplyr::bind_rows()
}
}
\references{
\url{https://www.fhwa.dot.gov/policyinformation/hpms.cfm}

\url{https://data-usdot.opendata.arcgis.com/datasets/usdot::highway-performance-monitoring-system-hpms-2020/about}
}
