% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urban_imperviousness.R
\name{get_urban_imperviousness}
\alias{get_urban_imperviousness}
\alias{install_urban_imperviousness}
\title{Get average buffered urban imperviousness}
\usage{
get_urban_imperviousness(x, year, buffer = 400)

install_urban_imperviousness(year = as.character(c(2021, 2019, 2016)))
}
\arguments{
\item{x}{a vector of s2 cell identifiers (\code{s2_cell} object)}

\item{year}{a character string that is the year of the urban imperviousness data}

\item{buffer}{distance from s2 cell (in meters) to summarize data}
}
\value{
for \code{get_urban_imperviousness()}, a numeric vector the same length as \code{x}

for \code{install_impervious()}, a character string path to impervious raster data
}
\description{
The percent urban imperviousness of each NLCD 30 x 30 m cell within \code{buffer} meters of
locations in \code{x} are averaged.
}
\details{
Urban imperviousness data is released sequentially, so the 2021 data is retrieved
from the 2023-06-30 release; both the 2019 and 2016 data are retrieved from the 2021-06-04 release.
From the \href{https://www.mrlc.gov/downloads/sciweb1/shared/mrlc/metadata/nlcd_2021_impervious_l48_20230630.xml}{metadata}
attribute definition for percent imperviousness: "while the file structure shows values in range from 0-255,
the values of 0-100 are the only real populated values, in addition to a background value of 127". Extracted values
greater than 100 are changed to 0 before averaging.
}
\examples{
get_urban_imperviousness(
  x = s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")),
  year = "2021"
)
}
\references{
\url{https://www.mrlc.gov/data/type/urban-imperviousness}

\url{https://www.mrlc.gov/viewer/}

\url{https://www.usgs.gov/centers/eros/science/national-land-cover-database}
}
