% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{install_elevation_data}
\alias{install_elevation_data}
\title{installs elevation data into user's data directory for the `appc` package}
\usage{
install_elevation_data()
}
\value{
path to elevation raster
}
\description{
installs elevation data into user's data directory for the `appc` package
}
