% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{install_geomarker_data}
\alias{install_geomarker_data}
\title{install geomarker data
Download, convert, and install all required geomarker data for \code{appc} predictions from 2016 - 2022. The smoke data depends on census tract data retrieved per state from the census API and is \emph{not} downloaded ahead of time here.}
\usage{
install_geomarker_data()
}
\value{
a character vector of the paths to installed geomarker data
}
\description{
install geomarker data
Download, convert, and install all required geomarker data for \code{appc} predictions from 2016 - 2022. The smoke data depends on census tract data retrieved per state from the census API and is \emph{not} downloaded ahead of time here.
}
