% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narr.R
\name{install_narr_data}
\alias{install_narr_data}
\title{installs NARR raster data into user's data directory for the \code{appc} package}
\usage{
install_narr_data(
  narr_var = c("air.2m", "hpbl", "acpcp", "rhum.2m", "vis", "pres.sfc", "uwnd.10m",
    "vwnd.10m"),
  narr_year = as.character(2016:2022)
)
}
\arguments{
\item{narr_var}{a character string that is the name of a NARR variable}

\item{narr_year}{a character string that is the year for the NARR data}
}
\value{
path to NARR raster data
}
\description{
installs NARR raster data into user's data directory for the \code{appc} package
}
\references{
https://psl.noaa.gov/data/gridded/data.narr.html
}
