% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_released_data.R
\name{install_released_data}
\alias{install_released_data}
\title{Download pre-installed data from GitHub release}
\usage{
install_released_data(
  released_data_name,
  package_version = packageVersion("appc")
)
}
\description{
\code{install_smoke_pm_data()}, \code{install_merra_data()},
\code{install_traffic()}, and install_nei_point_data()\verb{all download and then transform or subset data into smaller files to be used by appc}get_*_data()` functions.
Because this installation process can take a long time, the installed
geomarker data are (by default) downloaded from the
corresponding github release.
}
\details{
To turn \emph{off} the default usage of downloading
pre-generated data and to instead install data
geomarker data directly from their sources, set
options("appc_install_data_from_source"), or the
environment variable \code{APPC_INSTALL_DATA_FROM_SOURCE}
to any non-empty value.
}
\keyword{internal}
