% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke.R
\name{install_smoke_pm_data}
\alias{install_smoke_pm_data}
\title{Installs daily, census-tract level smoke pm data into user's data directory for the \code{appc} package}
\usage{
install_smoke_pm_data()
}
\value{
path to parquet file containing smoke data
}
\description{
See the examples to read the installed parquet file.
Merge this data with existing data on \code{date} and \code{census_tract_id_2010} to retrieve the
\code{smoke_pm} column. Note that any census tract-date combination implicitly missing has a value of zero.
}
\examples{
\dontrun{
arrow::read_parquet(install_smoke_pm_data())
}
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/36134580/}

\url{https://github.com/echolab-stanford/daily-10km-smokePM}
}
