% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_released_data.R
\name{install_source_preference}
\alias{install_source_preference}
\title{install_source_preference()}
\usage{
install_source_preference()
}
\description{
install_source_preference()
}
\keyword{internal}
