% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{install_traffic}
\alias{install_traffic}
\title{installs traffic data into user's data directory for the \code{appc} package}
\usage{
install_traffic()
}
\description{
installs traffic data into user's data directory for the \code{appc} package
}
\references{
https://www.fhwa.dot.gov/policyinformation/hpms.cfm

https://data-usdot.opendata.arcgis.com/datasets/usdot::highway-performance-monitoring-system-hpms-2020/about

https://www.fhwa.dot.gov/policyinformation/hpms/fieldmanual/hpms_field_manual_dec2016.pdf
}
