% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd.R
\name{install_treecanopy}
\alias{install_treecanopy}
\title{installs NLCD treecanopy raster data into user's data directory for the \code{appc} package}
\usage{
install_treecanopy(year = as.character(2021:2011))
}
\arguments{
\item{year}{a character string that is the year of the data}
}
\value{
path to treecanopy raster data
}
\description{
installs NLCD treecanopy raster data into user's data directory for the \code{appc} package
}
