% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_s2_cell.R
\name{latlon_to_s2_cell}
\alias{latlon_to_s2_cell}
\title{Convert latitude/longitude vectors into S2 cells}
\usage{
latlon_to_s2_cell(lat, lon)
}
\arguments{
\item{lat}{Numeric vector of latitudes (decimal degrees)}

\item{lon}{Numeric vector of longitudes (decimal degrees)}
}
\value{
An object of class \code{s2_cell}
}
\description{
Convert latitude/longitude vectors into S2 cells
}
\examples{
latlon_to_s2_cell(
  lat = c(45.0, 46.1),
  lon = c(-64.2, -65.3)
)
}
