% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_file}
\alias{pin_file}
\title{download and pin files}
\usage{
pin_file(url, overwrite = FALSE, progress = interactive())
}
\arguments{
\item{url}{URL of file to download}

\item{overwrite}{overwrite existing downloaded file with same name?}

\item{progress}{show messages about existing files and/or download progress bars?}
}
\value{
a character string that is the file path to the downloaded file (invisibly)
}
\description{
Files are downloaded to the R user data directory (i.e., \code{tools::R_user_dir("pin", "data")})
so they can be cached across all of an R user's sessions and projects.
Specify an alternative download location by setting the \code{R_USER_DATA_DIR} environment variable (see \code{?tools::R_user_dir}).
}
\examples{
\dontrun{
Sys.setenv("R_USER_DATA_DIR" = tempdir())
pin_file("https://geomarker.s3-us-east-2.amazonaws.com/nlcd_cog/nlcd_imperviousdesc_2019.tif")
}
}
