% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_pm25_date_range}
\alias{predict_pm25_date_range}
\title{Get daily PM2.5 model predictions using date ranges}
\usage{
predict_pm25_date_range(x, start_date, end_date, average = FALSE)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{start_date}{a date vector of start dates for each s2 cell, must be the same length as \code{x}}

\item{end_date}{a date vector of end dates for each s2 cell, must be the same length as \code{x}}

\item{average}{logical; summarize daily exposures estimates and standard errors?}
}
\description{
Get daily PM2.5 model predictions using date ranges
}
\details{
The standard error for averages of daily pm25 exposures with known standard errors
is calculated, assuming they are independent, as the square root of the sum of squared
individual standard errors divided the total number of individual daily pm25 exposures.
}
\examples{
predict_pm25_date_range(
  x = c("8841b39a7c46e25f", "8841a45555555555"),
  start_date = as.Date(c("2023-05-18", "2023-01-06")),
  end_date = as.Date(c("2023-06-22", "2023-08-15")),
  average = TRUE
)
}
