% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{contiguous_us}
\alias{contiguous_us}
\title{Get the geography of the 2020 contiguous United States}
\usage{
contiguous_us()
}
\value{
s2_geography object
}
\description{
Get the geography of the 2020 contiguous United States
}
\examples{
contiguous_us()
}
