% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{get_census_tract_id}
\alias{get_census_tract_id}
\title{Get census tract identifier of *the closest* census tract}
\usage{
get_census_tract_id(x, year)
}
\arguments{
\item{x}{a vector of s2 cell identifers (`s2_cell` object)}

\item{year}{a numeric data year passed to tigris to get state and tract boundaries}
}
\description{
Get census tract identifier of *the closest* census tract
}
\details{
`tigris::tracts()` powers this, so set `options(tigris_use_cache = TRUE)`
to benefit from its caching.
According to <https://github.com/walkerke/tigris>, available years for tracts
and states are 1990, 2000, 2010 - 2022
}
