% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_closest_year}
\alias{get_closest_year}
\title{Get the closest years to a vector of dates}
\usage{
get_closest_year(x, years = as.character(1800:2400))
}
\arguments{
\item{x}{a date vector}

\item{years}{vector of characters (or numerics) representing years to choose from}
}
\value{
a character vector of the closest year in \code{years} for each date in \code{x}
}
\description{
The time between a date and year is calculated using July 1st of the year.
}
\examples{
get_closest_year(x = as.Date(c("2021-05-15", "2022-09-01")), years = c("2020", "2022"))
}
