% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_closest_year}
\alias{get_closest_year}
\title{Get the closest years to a vector of dates}
\usage{
get_closest_year(date, years)
}
\arguments{
\item{date}{a vector of date objects}

\item{years}{vector of characters (or numerics) representing years to choose from}
}
\value{
a character vector of the closest year in \code{years} for each date in \code{date}
}
\description{
Get the closest years to a vector of dates
}
\details{
To find the closest year, each date is converted to a year
and the differences with the provided years is minimzed. This is a problem....
}
\examples{
\dontrun{
get_closest_year(as.Date(c("2021-09-15", "2022-09-01")), years = c(2020, 2022))
}
}
