% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_closest_year}
\alias{get_closest_year}
\title{get the closest years to a vector of dates}
\usage{
get_closest_year(date, years)
}
\arguments{
\item{date}{a vector of date objects}

\item{years}{vector of characters (or numerics) representing years to choose from}
}
\value{
a character vector of the closest year in \code{years} for each date in \code{date}
}
\description{
get the closest years to a vector of dates
}
\examples{
\dontrun{
get_closest_year(as.Date(c("2021-06-30", "2021-07-01")), years = 2021:2022)
}
}
