% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merra.R
\name{get_merra_data}
\alias{get_merra_data}
\alias{install_merra_data}
\alias{create_daily_merra_data}
\title{Get MERRA-2 aerosol diagnostics data}
\usage{
get_merra_data(x, dates)

install_merra_data(merra_year = as.character(2016:2023))

create_daily_merra_data(merra_date)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{dates}{a list of date vectors for the MERRA data, must be the same length as \code{x}}

\item{merra_year}{a character string that is the year for the merra data}

\item{merra_date}{a date object that is the date for the merra data}
}
\value{
for \code{get_merra_data()}, a list of tibbles the same
length as \code{x}, each containing merra data columns (...) with
one row per date in \code{dates}

for \code{install_merra_data()}, a character string path to the merra data

for \code{create_daily_merra_data()}, a tibble with columns for s2,
date, and concentrations of PM2.5 total, dust, oc, bc, ss, so4
}
\description{
Total and component (Dust, OC, BC, SS, SO4) surface PM2.5 concentrations
from the MERRA-2 \href{https://disc.gsfc.nasa.gov/datasets/M2T1NXAER_5.12.4/summary}{M2T1NXAER v5.12.4} product.
}
\details{
\itemize{
\item To install data from source, an \href{https://disc.gsfc.nasa.gov/information/documents?title=Data\%20Access}{Earthdata account linked with permissions for GES DISC} is required.
The \code{EARTHDATA_USERNAME} and \code{EARTHDATA_PASSWORD} must be set. If
a \code{.env} file is present, environment variables will be loaded
using the {dotenv} package.
\item Installed data are filtered to a
\href{http://bboxfinder.com/#24.766785,-126.474609,49.894634,-66.445313}{bounding box}
around the contiguous US, averaged to daily values, and
converted to micrograms per cubic meter ($ug/m^3$).
\item Total surface PM2.5 mass is calculated according to
the formula in \url{https://gmao.gsfc.nasa.gov/reanalysis/MERRA-2/FAQ/#Q4}
}
}
\examples{
# d <- list(
#   "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2023-11-06")),
#   "8841a45555555555" = as.Date(c("2023-06-22", "2023-08-15"))
# )
# get_merra_data(x = s2::as_s2_cell(names(d)), dates = d)
}
