% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd.R
\name{get_nlcd_summary}
\alias{get_nlcd_summary}
\title{get mean NLCD values buffered around a point}
\usage{
get_nlcd_summary(
  x,
  product = c("impervious", "treecanopy"),
  year,
  buffer = 400
)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{product}{get the "impervious" or "treecanopy" summary?}

\item{year}{a character string that is the year of the impervious or treecanopy data;
note that each product has different available years}

\item{buffer}{distance from s2 cell (in meters) to summarize data}
}
\value{
a vector of mean impervious or treecanopy values (the same length as \code{x})
}
\description{
get mean NLCD values buffered around a point
}
