% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urban_imperviousness.R
\name{install_urban_imperviousness}
\alias{install_urban_imperviousness}
\title{\code{install_urban_imperviousness()} installs NLCD urban imperviousness raster data into user's data directory for the \code{appc} package}
\usage{
install_urban_imperviousness(year = as.character(c(2021, 2019, 2016)))
}
\value{
for \code{install_impervious()}, a character string path to impervious raster data
}
\description{
\code{install_urban_imperviousness()} installs NLCD urban imperviousness raster data into user's data directory for the \code{appc} package
}
