% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{get_traffic_summary}
\alias{get_traffic_summary}
\title{get traffic summary data}
\usage{
get_traffic_summary(x, buffer = 400)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{buffer}{distance from s2 cell (in meters) to summarize data}
}
\value{
a list the same length as \code{x}, which each element having a list of \code{total_aadt_m} and \code{truck_aadt_m} estimates
}
\description{
get traffic summary data
}
\details{
A s2 level 15 approximation (~ 260 m sq) is used to simplify the intersection calculation with traffic summary data
}
\examples{
# get_traffic_summary(s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")), buffer = 1500)
}
