% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke.R
\name{install_smoke_pm_data}
\alias{install_smoke_pm_data}
\title{installs smoke pm data into user's data directory for the \code{appc} package}
\usage{
install_smoke_pm_data()
}
\value{
path to parquet file containing smoke data
}
\description{
note that any census tract - date combination implicitly missing has a value of zero
}
\examples{
\dontrun{
arrow::read_parquet(install_smoke_pm_data())
}
}
\references{
https://github.com/echolab-stanford/daily-10km-smokePM and https://pubmed.ncbi.nlm.nih.gov/36134580/
}
