% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_pm25}
\alias{predict_pm25}
\title{Get daily PM2.5 model predictions}
\usage{
predict_pm25(x, dates)
}
\arguments{
\item{x}{a vector of s2 cell identifers (\code{s2_cell} object)}

\item{dates}{a list of date vectors for the predictions, must be the same length as \code{x}}
}
\value{
a list of tibbles the same length as \code{x}, each containing
columns for the predicted (\code{pm25}) and its standard error (\code{pm25_se});
with one row per date in \code{dates}. These numerics are the concentrations of fine
particulate matter, measured in micrograms per cubic meter. See \code{vignette("cv-model-performance")}
for more details on the cross validated accuracy of the daily PM2.5 model predictions.
}
\description{
Get daily PM2.5 model predictions
}
\details{
Internally, loading the model file is cached, so repeated calls in the same R session
will not require the overhead of loading the model file for a new prediction.
}
\examples{
d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2023-11-06")),
  "8841a45555555555" = as.Date(c("2023-06-22", "2023-08-15"))
)

predict_pm25(x = s2::as_s2_cell(names(d)), dates = d)

# takes less time after called once because model file is cached in memory

d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-13", "2023-11-16")),
  "8841a45555555555" = as.Date(c("2023-06-21", "2023-08-25"))
)
predict_pm25(x = s2::as_s2_cell(names(d)), dates = d)
}
